+++ docs/guides/http_crawlers.mdx
@@ -24,12 +24,12 @@ It already includes almost all the functionality to crawl webpages and the only 
 
 ## Creating your own HTTP crawler.
 ### Why?
-In case you want to use some custom parser for parsing HTTP responses, and the rest of the <ApiLink to="class/AbstractHttpCrawler">`AbstractHttpCrawler`</ApiLink> functionality suit your needs.\
+In case you want to use some custom parser for parsing HTTP responses, and the rest of the <ApiLink to="class/AbstractHttpCrawler">`AbstractHttpCrawler`</ApiLink> functionality suit your needs.
 
 ### How?
-You need to define at least 2 new classes and decide what will be the type returned by the parser's `parse` method.\
-Parser will inherit from <ApiLink to="class/AbstractHttpParser">`AbstractHttpParser`</ApiLink> and it will need to implement all it's abstract methods.\
-Crawler will inherit from <ApiLink to="class/AbstractHttpCrawler">`AbstractHttpCrawler`</ApiLink> and it will need to implement all it's abstract methods.\
-Newly defined parser is then used in the `parser` argument of `AbstractHttpCrawler.__init__` method.\
+You need to define at least 2 new classes and decide what will be the type returned by the parser's `parse` method.
+Parser will inherit from <ApiLink to="class/AbstractHttpParser">`AbstractHttpParser`</ApiLink> and it will need to implement all it's abstract methods.
+Crawler will inherit from <ApiLink to="class/AbstractHttpCrawler">`AbstractHttpCrawler`</ApiLink> and it will need to implement all it's abstract methods.
+Newly defined parser is then used in the `parser` argument of `AbstractHttpCrawler.__init__` method.
 
 To get better idea and as an example please see one of our own HTTP-based crawlers mentioned above.
