+++ src/crawlee/abstract_http_crawler/__init__.py
@@ -1,11 +1,10 @@
 from ._abstract_http_crawler import AbstractHttpCrawler, HttpCrawlerOptions
-from ._abstract_http_parser import AbstractHttpParser, BlockedInfo
+from ._abstract_http_parser import AbstractHttpParser
 from ._http_crawling_context import ParsedHttpCrawlingContext
 
 __all__ = [
     'AbstractHttpCrawler',
     'AbstractHttpParser',
-    'BlockedInfo',
     'HttpCrawlerOptions',
     'ParsedHttpCrawlingContext',
 ]
