+++ src/crawlee/basic_crawler/__init__.py
@@ -1,6 +1,6 @@
 from crawlee._types import BasicCrawlingContext
 
-from ._basic_crawler import BasicCrawler, BasicCrawlerOptions
+from ._basic_crawler import BasicCrawler, BasicCrawlerOptions, BlockedInfo
 from ._context_pipeline import ContextPipeline
 
-__all__ = ['BasicCrawler', 'BasicCrawlerOptions', 'BasicCrawlingContext', 'ContextPipeline']
+__all__ = ['BasicCrawler', 'BasicCrawlerOptions', 'BasicCrawlingContext', 'BlockedInfo', 'ContextPipeline']
